+++ data-remote/src/main/java/com/data_remote/api/TagApiService.kt
@@ -2,8 +2,12 @@ package com.data_remote.api
 
 import retrofit2.http.GET
 
-interface HomeApiService {
+interface TagApiService {
     @GET("/tag/gettag")
     suspend fun getTag(
+    ):List<String>
+
+    @GET("/tag/popularpost")
+    suspend fun getPopularTag(
     ):List<String>
 }
